Allow <Automatic> mode syncs between 32 bit, 64 bit, Windows and Linux builds Show progress indicator in window title Support for progress indicator in Windows 7 Superbar Reduced progress indicator flicker Prevent silent batch mode from taking keyboard focus Improved error messages (loading/saving/copying files) Improved environment variable tolerance: strip blanks and double-quotes RealTimeSync: Fixed crash when double-clicking systray icon Allow aborting all operations via Escape key Added British English translation

+++ library/CustomGrid.h
@@ -65,7 +65,7 @@ public:
                       CustomGridRight*  gridRight,
                       const FreeFileSync::GridView* gridDataView);
 
-    std::set<unsigned int> getAllSelectedRows() const;
+    std::set<size_t> getAllSelectedRows() const;
 
     //set sort direction indicator on UI
     typedef int SortColumn;
